+++ packages/tailwindcss-language-server/src/language/cssServer.ts
@@ -334,6 +334,12 @@ function replace(delta = 0) {
     return `@media(${MEDIA_MARKER})${' '.repeat(p1.length + delta)}{`
   }
 }
+function replaceWithStyleRule(delta = 0) {
+  return (_match: string, p1: string) => {
+    let spaces = ' '.repeat(p1.length + delta)
+    return `.foo${spaces}{`
+  }
+}
 
 function createVirtualCssDocument(textDocument: TextDocument): TextDocument {
   let content = textDocument
@@ -347,6 +353,9 @@ function createVirtualCssDocument(textDocument: TextDocument): TextDocument {
     .replace(/@screen(\s+[^{]+){/g, replace(-2))
     .replace(/@variants(\s+[^{]+){/g, replace())
     .replace(/@responsive(\s*){/g, replace())
+    .replace(/@utility(\s+[^{]+){/g, replaceWithStyleRule())
+    .replace(/@custom-variant(\s+[^{]+){/g, replaceWithStyleRule())
+    .replace(/@variant(\s+[^{]+){/g, replaceWithStyleRule())
     .replace(/@layer(\s+[^{]{2,}){/g, replace(-3))
     .replace(/@reference\s*([^;]{2,})/g, '@import    $1')
     .replace(
@@ -401,7 +410,7 @@ async function validateTextDocument(textDocument: TextDocument): Promise<void> {
     .filter((diagnostic) => {
       if (
         diagnostic.code === 'unknownAtRules' &&
-        /Unknown at rule @(tailwind|apply|config|theme|plugin|source|utility|variant|custom-variant)/.test(
+        /Unknown at rule @(tailwind|apply|config|theme|plugin|source|utility|variant|custom-variant|slot)/.test(
           diagnostic.message,
         )
       ) {