+++ packages/tailwindcss-language-server/src/project-locator.ts
@@ -16,6 +16,7 @@ import { extractSourceDirectives, resolveCssImports } from './css'
 import { normalizeDriveLetter, normalizePath, pathToFileURL } from './utils'
 import postcss from 'postcss'
 import * as oxide from './oxide'
+import { guessTailwindVersion, TailwindVersion } from './version-guesser'
 
 export interface ProjectConfig {
   /** The folder that contains the project */
@@ -130,6 +131,8 @@ export class ProjectLocator {
   private async createProject(config: ConfigEntry): Promise<ProjectConfig | null> {
     let tailwind = await this.detectTailwindVersion(config)
 
+    let possibleVersions = config.entries.flatMap((entry) => entry.versions)
+
     console.log(
       JSON.stringify({
         tailwind,
@@ -157,6 +160,15 @@ export class ProjectLocator {
     if (config.type === 'css') {
       // This version of Tailwind doesn't support CSS-based configuration
       if (!tailwind.features.includes('css-at-theme')) {
+        return null
+      }
+
+      // This config doesn't include any v4 features (even ones that were also in v3)
+      if (!possibleVersions.includes('4')) {
+        console.warn(
+          `The config ${config.path} looks like it might be for a different Tailwind CSS version. Skipping.`,
+        )
+
         return null
       }
 
@@ -324,6 +336,9 @@ export class ProjectLocator {
     // Read the content of all the CSS files
     await Promise.all(css.map((entry) => entry.read()))
 
+    // Determine what tailwind versions each file might be using
+    await Promise.all(css.map((entry) => entry.resolvePossibleVersions()))
+
     // Keep track of files that might import or involve Tailwind in some way
     let imports: FileEntry[] = []
 
@@ -331,8 +346,9 @@ export class ProjectLocator {
       // If the CSS file couldn't be read for some reason, skip it
       if (!file.content) continue
 
-      // Look for `@import`, `@tailwind`, `@theme`, `@config`, etc…
-      if (!file.isMaybeTailwindRelated()) continue
+      // This file doesn't appear to use Tailwind CSS nor any imports
+      // so we can skip it
+      if (file.versions.length === 0) continue
 
       // Find `@config` directives in CSS files and resolve them to the actual
       // config file that they point to. This is only relevant for v3 which
@@ -642,6 +658,7 @@ class FileEntry {
   deps: FileEntry[] = []
   realpath: string | null
   sources: string[] = []
+  versions: TailwindVersion[] = []
 
   constructor(
     public type: 'js' | 'css',
@@ -710,6 +727,13 @@ class FileEntry {
   }
 
   /**
+   * Determine which Tailwind versions this file might be using
+   */
+  async resolvePossibleVersions() {
+    this.versions = this.content ? guessTailwindVersion(this.content) : []
+  }
+
+  /**
    * Look for `@config` directives in a CSS file and return the path to the config
    * file that it points to. This path is (possibly) relative to the CSS file so
    * it must be resolved to an absolute path before returning.
@@ -726,25 +750,6 @@ class FileEntry {
     }
 
     return normalizePath(path.resolve(path.dirname(this.path), match.groups.config.slice(1, -1)))
-  }
-
-  /**
-   * Look for tailwind-specific directives in a CSS file. This means that it
-   * participates in the CSS "graph" for the project and we need to traverse
-   * the graph to find all the CSS files that are considered entrypoints.
-   */
-  isMaybeTailwindRelated(): boolean {
-    if (!this.content) return false
-
-    let HAS_IMPORT = /@import\s*['"]/
-    let HAS_TAILWIND = /@tailwind\s*[^;]+;/
-    let HAS_DIRECTIVE = /@(theme|plugin|config|utility|variant|apply)\s*[^;{]+[;{]/
-
-    return (
-      HAS_IMPORT.test(this.content) ||
-      HAS_TAILWIND.test(this.content) ||
-      HAS_DIRECTIVE.test(this.content)
-    )
   }
 }